+++ scripts/embedding/swipl/swipl_logtalk_qlf.ps1
@@ -8,8 +8,8 @@
 ##   Last updated on March 18, 2025
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
-##   SPDX-FileCopyrightText: 2022 Hans N. Beck
-##   SPDX-FileCopyrightText: 2022 Paulo Moura <pmoura@logtalk.org>
+##   Copyright 2022-2025 Paulo Moura <pmoura@logtalk.org>
+##   Copyright 2022 Hans N. Beck
 ##   SPDX-License-Identifier: Apache-2.0
 ##   
 ##   Licensed under the Apache License, Version 2.0 (the "License");