+++ tools/lgtdoc/xml/lgt2rst.ps1
@@ -4,7 +4,8 @@
 ##   Last updated on March 21, 2025
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
-##   Copyright 2022 Hans N. Beck and Paulo Moura <pmoura@logtalk.org>
+##   Copyright 2022-2025 Paulo Moura <pmoura@logtalk.org>
+##   Copyright 2022 Hans N. Beck
 ##   SPDX-License-Identifier: Apache-2.0
 ##   
 ##   Licensed under the Apache License, Version 2.0 (the "License");