+++ RELEASE_NOTES.md
@@ -74,7 +74,8 @@ integration scripts to a module.
 * IMPROVED: The `logtalk_tester.*` scripts handling of the `-b` and `-u`
 options.
 
-* IMPROVED: The `lgtenv.ps1` PowerShell script handling of file paths.
+* IMPROVED: The `lgt2svg.ps1` and `lgtenv.ps1` PowerShell scripts handling
+of file paths.
 
 * FIXED: The `logtalk_tester.ps1` PowerShell script `xunit_net_v2` format
 option.